Distribute tests from DataFrameSuite to more specific suites
