+++ sql/core/src/test/scala/org/apache/spark/sql/DataFrameJoinSuite.scala
@@ -602,4 +602,14 @@ class DataFrameJoinSuite extends QueryTest
       )
     }
   }
+
+  test("SPARK-20359: catalyst outer join optimization should not throw npe") {
+    val df1 = Seq("a", "b", "c").toDF("x")
+      .withColumn("y", udf{ (x: String) => x.substring(0, 1) + "!" }.apply($"x"))
+    val df2 = Seq("a", "b").toDF("x1")
+    df1
+      .join(df2, df1("x") === df2("x1"), "left_outer")
+      .filter($"x1".isNotNull || !$"y".isin("a!"))
+      .count()
+  }
 }
